+++ src/pages/creator/layout/CreatorLayout.vue
@@ -24,7 +24,7 @@ export default {
   },
   data () {
     return {
-      collapse: true
+      collapse: false
     }
   },
   created () {